rename: 지원하기 모달 이름 수정
