+++ src/pages/PotDetail/components/ApplyModal/ApplyModal.style.ts
